requirements.txt
openai
python-dotenv
termcolor  # makes the terminal output colorful and easier to read